// src/Typings/OptionSets/IShareDoOptionSet.ts
export interface IShareDoOptionSet {
  children: any[];
  id: number;
  optionSetName: string;
  parentValueId?: any;
  name: string;
  shortName?: any;
  iconClass?: any;
  isActive: boolean;
  meaningCode?: any;
  migrationId?: any;
  displayColour: string;
  deleted: boolean;
  defaultActive: boolean;
  displayOrder: number;
}